prisma 2 db seed works
