+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3'
 
 services:
-  node:
+  app:
     build:
       context: .
       args:
@@ -36,37 +36,47 @@ services:
       - APP_SECRET=secret1234
       - EMAIL_FROM=no-reply@schooledlunch.club
     depends_on:
-      - prisma
+      - database
     healthcheck:
       disable: true
 
-  prisma:
-    image: prismagraphql/prisma:1.34
-    restart: always
-    ports:
-      - '4466:4466'
-    environment:
-      PRISMA_CONFIG: |
-        port: 4466
-        databases:
-          default:
-            connector: mysql
-            host: mysql
-            port: 3306
-            user: root
-            password: prisma
-            migrations: true
+  # prisma:
+  #   image: prismagraphql/prisma:1.34
+  #   restart: always
+  #   ports:
+  #     - '4466:4466'
+  #   environment:
+  #     PRISMA_CONFIG: |
+  #       port: 4466
+  #       databases:
+  #         default:
+  #           connector: mysql
+  #           host: mysql
+  #           port: 3306
+  #           user: root
+  #           password: prisma
+  #           migrations: true
 
-  mysql:
-    image: mysql:5.7
+  database:
+    image: postgres:12
     restart: always
     environment:
-      MYSQL_ROOT_PASSWORD: prisma
+      POSTGRES_PASSWORD: secretsauce
     ports:
-      - '3306:3306'
+      - '5433:5432'
     volumes:
-      - mysql:/var/lib/mysql
+      - pgdata:/var/lib/postgresql/data
+
+  # mysql:
+  #   image: mysql:5.7
+  #   restart: always
+  #   environment:
+  #     MYSQL_ROOT_PASSWORD: prisma
+  #   ports:
+  #     - '3306:3306'
+  #   volumes:
+  #     - mysql:/var/lib/mysql
 
 volumes:
   notused:
-  mysql:
+  pgdata:
