+++ package.json
@@ -22,14 +22,12 @@
     "lint-js": "eslint \"**/*.{js,jsx}\"",
     "prettier": "prettier --write \"**/*.{js,jsx,json,yml,xml,md,css}\"",
     "build": "next build",
-    "prisma-deploy": "prisma deploy",
-    "prisma-generate": "prisma generate",
-    "prisma-seed": "prisma seed",
-    "prisma-reset": "prisma reset"
+    "seed": "node server/seed.js"
   },
   "dependencies": {
     "@apollo/react-hooks": "3.1.5",
     "@apollo/react-ssr": "3.1.5",
+    "@prisma/client": "2.1.3",
     "apollo-cache-inmemory": "1.6.6",
     "apollo-client": "2.6.10",
     "apollo-link-context": "1.0.20",
@@ -71,6 +69,7 @@
   },
   "devDependencies": {
     "@babel/core": "7.10.2",
+    "@prisma/cli": "2.1.3",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.1.0",
     "babel-jest": "26.0.1",
@@ -91,7 +90,6 @@
     "postcss-calc": "7.0.2",
     "postcss-preset-env": "6.7.0",
     "prettier": "2.0.5",
-    "prisma": "1.34.10",
     "react-addons-test-utils": "15.6.2",
     "react-test-renderer": "16.13.1",
     "stylelint": "13.6.0",
