+++ prisma/.env
@@ -1,0 +1,9 @@
+# Environment variables declared in this file are automatically made available to Prisma.
+# See the documentation for more detail: https://pris.ly/d/prisma-schema#using-environment-variables
+
+# Prisma supports the native connection string format for PostgreSQL, MySQL and SQLite.
+# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
+
+# DATABASE_URL="mysql://root:prisma@mysql:3306/default@default"
+
+DATABASE_URL="postgresql://postgres:secretsauce@database:5432/recipelab"
